tools: Add early-scratch-setup initscript
- Add early-scratch-setup initscript for Live CD, so named because it
needs to run before livesys, which also starts at sequence number 0.
The script uses gather_free_space to create an encrypted scratch
volume out of free space on the local disk, creates and enables a swap
partition inside it, and creates and mounts /home from the remainder.
The script only runs if the scratch_volume option is specified on the
kernel command line.  Live-specific devices are excluded from the
scratch volume.