iso: Enable nscd
- glibc caches failed DNS lookups and does not retry them when a DNS
server becomes available.  As a result, when NetworkManager first finds
a network connection, existing processes are unable to use it correctly
unless they call res_init() to reset the cache.  This is possible to do
from Python, but not straightforward.  The glibc maintainer claims that
the proper solution is for the system to use nscd (see Red Hat #442172),
which effectively makes flushing the cache nscd's problem rather than
the application's.  Therefore, enable nscd.