Make sure /mnt/live is writable by liveuser in normal mode
- tools: Add remount_live_volume script to do the processing necessary
to remount /mnt/live in normal mode.  (Remounting is not possible in
persistent mode.)  If /dev/live is VFAT, ensure it is mounted writable
by liveuser.  If it is ext[234], ensure it is mounted.  If it is a CD
filesystem, ensure it is *un*mounted, since there's no value to leaving
it mounted and keeping the drive door locked.

- iso: Replace iso9660 unmount special case in livesys initscript with a
call to remount_live_volume