tools/pocket_isr_update: Don't show update buttons if booted from CD
- Don't show the update icon in the taskbar or the update button in
the notification popup if we're booted from CD.  In the other cases where
we can't update, the user can fix the impediment, so it's useful to
explain the problem to them.  When booted from CD the user already
understands that we can't update and why, so don't show misleading UI
widgets.