tools/pocket_isr_update: Don't be confused by captive portals
We were assuming that if the data at VERSION_URL is different from the
current Pocket ISR version, we need to update.  This fails in the
presence of a captive portal that redirects the HTTP fetch to a
different page.

- Fetch update metadata from a different URL containing key-value pairs,
enabling us to check whether we've received proper data or not.  Expect
a Version key containing the available Pocket ISR version.

- If we fail to parse the update metadata, assume we're behind a captive
portal and try again in 5 minutes (in case the user has manually
activated the portal)

- Refuse to update if the available version is older than our current
version.  This prevents a malicious proxy from forcing us to downgrade
to a signed but older image containing a security flaw.

- Perform all _query() error handling in exception handlers