tools/gather_free_space: Add -r/--report option
- Add -r/--report option to print a results summary in (hand-coded)
YAML format to the specified file, or to stdout if "-" is specified.
Write the report only when we're done, so we never leave a partial
report file; any fatal error will cause the report not to be written.
For now, only report device rejections.

- Don't immediately die() on minsize violation; instead, still write
the report and clean up data structures

- For simplicity, print final results notice immediately after running
the DM task rather than after we clean up

- Add and use warn() macro for non-fatal errors